add product images part

+++ src/main/java/com/don8/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.don8.repository;
 
-import com.don8.model.Product;
+import com.don8.model.dbentity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.Optional;
@@ -9,6 +9,6 @@ import java.util.Optional;
  @author BHARAT VAYITLA
  */
 
-public interface ProductRepository extends JpaRepository<Product, BigInteger> {
-    Optional<Product> findByPid(BigInteger pid);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Optional<Product> findByPid(Long pid);
 }
